+++ README.md
@@ -1,7 +1,7 @@
 # Introduction
 
 This is a project to build a simple web application to interact with and visualise NASA's ENONET [API](https://eonet.sci.gsfc.nasa.gov/).
-The data from this API provides information about past and current worldwide natural events. For simplicity, the application avoids any JavaScript frameworks and explores some of the newer features in ES2015 and beyond. Bootstrap is used for styling and  chart.js and mapbox are used for visualisation.  
+The data from this API provides information about past and current worldwide natural events. For simplicity, the application avoids any JavaScript frameworks and explores some of the newer features in ES6 and beyond. Bootstrap is used for styling and chart.js / mapbox are used for visualisation.
