+++ charts.html
@@ -7,7 +7,7 @@
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
     integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.0/font/bootstrap-icons.css">
-  <link rel='stylesheet' type='text/css' media='screen' href='styles.css'>
+  <link rel='stylesheet' type='text/css' media='screen' href='/styles/styles.css'>
 </head>
 
 <body>
@@ -71,12 +71,7 @@
     integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
     crossorigin="anonymous"></script>
   <script src="https://cdn.jsdelivr.net/npm/chart.js@3.6.2/dist/chart.min.js"></script>
-  <script src="eonet-api.js"></script>
-  <script>
-    window.addEventListener('load', (event) => {
-      initChartControls();
-    });
-  </script>
+  <script type="module" src="/scripts/charts.js"></script>
 </body>
 
 </html>
