+++ modules/common.js
@@ -1,0 +1,2 @@
+export const EONET_BASE_URL = "https://eonet.gsfc.nasa.gov/api/v3";
+
