+++ scripts/charts.js
@@ -1,0 +1,6 @@
+
+import { initChartControls  } from '../modules/charts.js';
+
+window.addEventListener('load', (event) => {
+    initChartControls();
+  });
