+++ scripts/dashboard.js
@@ -1,0 +1,6 @@
+
+import { fetchDashboardDataAsync } from '../modules/events.js';
+
+window.addEventListener('load', (event) => {
+    fetchDashboardDataAsync();
+});
