+++ scripts/index.js
@@ -1,0 +1,5 @@
+import { fetchCategoriesAsync } from '../modules/categories.js';
+
+window.addEventListener('load', (event) => {
+    fetchCategoriesAsync();
+});
