+++ scripts/maps.js
@@ -1,0 +1,6 @@
+import { fetchMapDataAsync  } from '../modules/maps.js';
+
+window.addEventListener('load', (event) => {
+    fetchMapDataAsync();
+});
+
